feat(settings): migrate settings page to server/client architecture
- Convert settings page from client-only to hybrid server/client components
- Add server component for static header and metadata (SEO optimization)
- Extract dynamic functionality to SettingsContent client component
- Create SettingsSkeleton for loading states with Suspense streaming
- Add /api/settings route with GET/PUT endpoints
- Implement updateSettingsForCurrentUser in server data fetching utilities
- Maintain all existing functionality: settings management, data export/import, clear data
- Preserve localStorage as fallback during transition period
- Follow established Next.js 14 App Router patterns for performance
